REFACTOR: move epoch layout conversions to data helpers
