+++ mne_denoise/sns/core.py
@@ -28,13 +28,14 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
 from .._covariance import compute_covariance
-from .._data import extract_data_from_mne, reconstruct_mne_object
-from .._logging import set_log_level_from_verbose
-from .._spatial import (
-    apply_spatial_transform,
+from .._data import (
     continuous_to_epochs,
     epochs_to_continuous,
+    extract_data_from_mne,
+    reconstruct_mne_object,
 )
+from .._logging import set_log_level_from_verbose
+from .._spatial import apply_spatial_transform
 from .._validation import (
     check_channel_first_data,
     check_channel_layout,
