+++ mne_denoise/zapline/core.py
@@ -41,13 +41,14 @@ import warnings
 
 import numpy as np
 
-from .._data import extract_data_from_mne, reconstruct_mne_object
-from .._logging import set_log_level_from_verbose
-from .._spatial import (
-    apply_spatial_transform,
+from .._data import (
     continuous_to_epochs,
     epochs_to_continuous,
+    extract_data_from_mne,
+    reconstruct_mne_object,
 )
+from .._logging import set_log_level_from_verbose
+from .._spatial import apply_spatial_transform
 
 # Inherit from DSS
 from ..dss.denoisers.spectral import LineNoiseBias
